requirements.txt
# Unity MCP Server Requirements

# MCP Protocol
mcp>=0.9.0

# Web Scraping
playwright>=1.56.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector Database
chromadb>=0.4.0

# AI/ML
openai>=1.0.0

# Async HTTP
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0
